Normalized the distance before calculating the reward
In _REINFORCEjs: https://github.com/karpathy/reinforcejs, the distances are normalized when it calculates the reward. However, the reward in puckworld.py is calculated according to the absolute distance to the good creep but the normalized distance to the bed creep. It causes a problem that the player always not sensitive to the approaching of the bed creep. Now the code is revised to calculate the normalized distance for both when is calculates the reward.